找到了long long型出错的原因，完成了一构造函数、插入结点函数、输出重载的设计与试

+++ Hugeint.cpp
@@ -15,7 +15,7 @@ Hugeint::Hugeint(long long int num, unsigned int rank)//ToDo
 		sign=-1;
 		num=-num;
 	}
-	int h1,h0;
+	long long h1,h0;
 	h0=num%B;
 	h1=num/B;
 	if(h0) Insert(h0,rank);
